+++ lucene/core/src/java/org/apache/lucene/util/bkd/BKDWriter.java
@@ -1441,8 +1441,8 @@ public class BKDWriter implements Closeable {
         heapSource = (HeapPointWriter) points.writer;
       }
 
-      int from = (int) points.start;
-      int to = (int) (points.start + points.count);
+      int from = Math.toIntExact(points.start);
+      int to = Math.toIntExact(points.start + points.count);
       //we store common prefix on scratch1
       computeCommonPrefixLength(heapSource, scratch1, from, to);
